+++ demo/demo.py
@@ -381,21 +381,20 @@ if __name__ == "__main__":
             "name": "SevenBandParametricEQ",
         },
         {
-            "instance": Shift(min_fraction=-0.5, max_fraction=0.5, fade=False, p=1.0),
+            "instance": Shift(min_shift=-0.5, max_shift=0.5, fade_duration=0.0, p=1.0),
             "num_runs": 5,
             "name": "ShiftWithoutFade",
         },
         {
-            "instance": Shift(min_fraction=-0.5, max_fraction=0.5, fade=True, p=1.0),
+            "instance": Shift(min_shift=-0.5, max_shift=0.5, fade_duration=0.01, p=1.0),
             "num_runs": 5,
             "name": "ShiftWithShortFade",
         },
         {
             "instance": Shift(
-                min_fraction=-0.5,
-                max_fraction=0.5,
+                min_shift=-0.5,
+                max_shift=0.5,
                 rollover=False,
-                fade=True,
                 fade_duration=0.3,
                 p=1.0,
             ),
@@ -417,7 +416,7 @@ if __name__ == "__main__":
                             PitchShift(min_semitones=-4, max_semitones=4, p=1.0),
                         ],
                     ),
-                    Shift(min_fraction=-0.5, max_fraction=0.5, p=0.5),
+                    Shift(min_shift=-0.5, max_shift=0.5, p=0.5),
                     OneOf([TanhDistortion(p=1.0), ClippingDistortion(p=1.0)], p=0.25),
                 ]
             ),